modules/location-service/location-consumer/requirements.txt
kafka-python
grpcio==1.67.0
grpcio-tools==1.67.0
psycopg2-binary
